requirements.txt
opencv-python
pillow
tk
paho-mqtt==1.5.1
